pyproject.toml
[project]
name = "speech-model"
version = "0.0.0"
description = "Speech classification model"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "transformers>=4.30.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.66.0",
    "wandb>=0.16.0",
    "pyyaml>=6.0",
    "numpy>=2.2.6",
    "google-genai>=1.56.0",
]

[build-system]
requires = ["uv_build>=0.9.12,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
]
data-acquisition = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "tqdm>=4.66.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "soundfile>=0.12.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

